src/requirements.txt
schedule
azure-iot-device